move MST pretty print code to new file
